% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{n_exceedances}
\alias{n_exceedances}
\title{Extract the number of exceedances from a fitted model}
\usage{
n_exceedances(fit, y)
}
\arguments{
\item{fit}{An \code{nhpp_fit} object.}

\item{y}{Numeric vector of observations used to fit the model.}
}
\value{
Integer. Number of observations exceeding the threshold.
}
\description{
Extract the number of exceedances from a fitted model
}
